requirements.txt
anyio
certifi
charset-normalizer
click
databases
dnspython
email-validator
fastapi
fastapi-utils
greenlet
h11
httpcore
httptools
httpx
idna
itsdangerous
Jinja2
MarkupSafe
orjson
psycopg2-binary
pydantic
python-dotenv
python-multipart
PyYAML
requests
rfc3986
sniffio
SQLAlchemy
starlette
typing
typing_extensions
ujson
urllib3
uvicorn
watchfiles
websockets
pandas
openai
langchain
cdifflib
unidecode
pypdf
python-pptx
tiktoken